chrome-extension/server/requirements.txt
fastapi
uvicorn
litellm
playwright
pydantic
python-dotenv
